Fix lightmap UV-mapping issues in shib8

+++ Quake/gl_model.h
@@ -161,7 +161,7 @@ typedef struct msurface_s
 	int			firstedge;	// look up in model->surfedges[], negative numbers
 	int			numedges;	// are backwards edges
 
-	short		texturemins[2];
+	int			texturemins[2];
 	short		extents[2];
 
 	int			light_s, light_t;	// gl lightmap coordinates
